backend/composite_services/refund_ticket/requirements.txt
Flask==3.1.0
Flask-Cors==3.0.10
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.1.3
pika==1.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
urllib3==2.3.0
Werkzeug==3.1.3
flask-cors==4.0.0
pika==1.3.2
